Refactors node color assignment and adjacency map creation; moves utility functions to utils.ts and removes unused node-collapse.ts
