+++ components/react-flow-nodes.tsx
@@ -1,68 +1,13 @@
-import { getDescendants } from "@/core/node-collapse";
 import { ReactFlowNode } from "@/core/react-flow.service";
 import { useGraphStore } from "@/lib/graphStore";
+import { getDescendants, getNodeColors } from "@/lib/utils";
 import { Handle, Position, useReactFlow } from "@xyflow/react";
 import {
 	LucideMinus,
 	LucidePlus
 } from "lucide-react";
 
-const getNodeColors = (nodeType: string) => {
-  switch (nodeType) {
-    case "class":
-    case "interface":
-      return {
-        from: "#7c3aed", // purple-600
-        to: "#5b21b6", // purple-800
-        border: "#a78bfa", // purple-400
-        handle: "#ddd6fe", // purple-200
-        text: "#e9d5ff", // purple-200
-      };
-    case "method":
-    case "function":
-    case "constructor":
-      return {
-        from: "#f59e0b", // amber-500
-        to: "#d97706", // amber-600
-        border: "#fbbf24", // amber-400
-        handle: "#fde68a", // amber-200
-        text: "#fde68a", // amber-200
-      };
-    case "variable":
-    case "property":
-      return {
-        from: "#ef4444", // red-500
-        to: "#dc2626", // red-600
-        border: "#f87171", // red-400
-        handle: "#fecaca", // red-200
-        text: "#fecaca", // red-200
-      };
-    case "file":
-      return {
-        from: "#10b981", // emerald-500
-        to: "#059669", // emerald-600
-        border: "#34d399", // emerald-400
-        handle: "#a7f3d0", // emerald-200
-        text: "#a7f3d0", // emerald-200
-      };
-    case "folder":
-      return {
-        from: "#0ea5e9", // sky-500
-        to: "#0284c7", // sky-600
-        border: "#38bdf8", // sky-400
-        handle: "#bae6fd", // sky-200
-        text: "#bae6fd", // sky-200
-      };
-    default:
-      return {
-        from: "#6b7280", // gray-500
-        to: "#4b5563", // gray-600
-        border: "#9ca3af", // gray-400
-        handle: "#d1d5db", // gray-300
-        text: "#d1d5db", // gray-300
-      };
-  }
-};
+
 
 interface BaseNodeProps {
   data: ReactFlowNode["data"];
@@ -274,53 +219,6 @@ export function GraphNode({ data, selected }: BaseNodeProps) {
   );
 }
 
-export function CompactNode({ data, selected }: BaseNodeProps) {
-  // Get colors based on node type - matching formatService.ts
-
-  const { getNode, getNodeConnections, addEdges, addNodes, deleteElements } =
-    useReactFlow();
-  const colors = getNodeColors(data.type);
-
-  return (
-    <div
-      className="border-2 rounded-lg p-3 shadow-lg bg-cyan-100 hover:shadow-xl transition-all cursor-pointer min-w-[160px] max-w-[220px] w-[220px] overflow-hidden"
-      onClick={() => {
-        console.log("Compact node clicked", data.id);
-        const fullNode = getNode(data.id.replace("-compact", ""));
-        if (fullNode) {
-          const connections = getNodeConnections({ nodeId: fullNode.id });
-          deleteElements({
-            edges: connections.map((c) => ({ id: c.edgeId })),
-          });
-          addNodes([fullNode]);
-          addEdges(
-            connections.map((c) => ({
-              id: c.edgeId,
-              source: c.source,
-              target: c.target,
-            }))
-          );
-        }
-      }}
-    >
-      <Handle
-        type="target"
-        position={Position.Top}
-        className="w-3 h-3"
-        style={{ backgroundColor: colors.handle }}
-      />
-
-      <div className="flex items-center gap-2">{data.label}</div>
-
-      <Handle
-        type="source"
-        position={Position.Bottom}
-        className="w-3 h-3"
-        style={{ backgroundColor: colors.handle }}
-      />
-    </div>
-  );
-}
 
 export function StubNode({ data, selected }: BaseNodeProps) {
   // Get colors based on node type - matching formatService.ts
@@ -374,5 +272,4 @@ export const nodeTypes = {
   c2SubcategoryNode: C2SubcategoryNode,
   node: GraphNode,
   stubNode: StubNode,
-  compactNode: CompactNode,
 };
